Started big BingoTask refactor, making it not abstract anymore and using records for specific task data to prepare for statistics

+++ src/main/java/io/github/steaf23/bingoreloaded/util/CountdownTimer.java
@@ -55,13 +55,13 @@ public class CountdownTimer extends GameTimer
     }
 
     @Override
-    public int pause()
+    public long pause()
     {
         return getTime();
     }
 
     @Override
-    public int stop()
+    public long stop()
     {
         try
         {
