+++ src/main/java/io/github/steaf23/bingoreloaded/util/CounterTimer.java
@@ -31,13 +31,13 @@ public class CounterTimer extends GameTimer
     }
 
     @Override
-    public int pause()
+    public long pause()
     {
         return getTime();
     }
 
     @Override
-    public int stop()
+    public long stop()
     {
         try
         {
