+++ src/main/java/io/github/steaf23/bingoreloaded/util/TimeNotifier.java
@@ -2,5 +2,5 @@ package io.github.steaf23.bingoreloaded.util;
 
 public interface TimeNotifier
 {
-    public abstract void timeUpdated(int seconds);
+    public abstract void timeUpdated(long seconds);
 }
